refactor(live): move cli interface to separate file
