+++ javascript-modules/live/package.json
@@ -7,10 +7,10 @@
   "main": "main.js",
   "scripts": {
     "test": "nyc ava -v",
-    "manual-test": "node main.js -b .test/fixtures -o .test/manual/index.js && cd .test/manual && python -m SimpleHTTPServer"
+    "manual-test": "node cli.js -b .test/fixtures -o .test/manual/index.js && cd .test/manual && python -m SimpleHTTPServer"
   },
   "bin": {
-    "bookshop-live": "main.js"
+    "bookshop-live": "cli.js"
   },
   "author": "@bglw",
   "license": "MIT",
